version 1.2.2

+++ src/ibm2ieee.c
@@ -73,7 +73,7 @@ void ibm2ieee(register unsigned char *out, register const unsigned char *in, int
 	 */
 	register int	i;
 	for( i=count-1; i >= 0; i-- )  {
-		register unsigned long left, right, signbit;
+		register unsigned int left, right, signbit;
 		register int exp;
 
 		left  = (in[0]<<24) | (in[1]<<16) | (in[2]<<8) | in[3];
